refactor: extract @rn-iso/core -- the shared primitives, implemented once
The three deliberate copies of the cache-root resolution, buildCacheKey
and the caches.json self-registration collapse into a tiny CJS hard
dependency of all three packages. The old prohibition was on importing
RN-ISO ITSELF (optional ESM peer, silently absent under npx); a plain
always-installed CJS dep has neither failure mode. Public APIs unchanged;
cache-packages.test keeps pinning cross-package agreement and core's own
isolation; release flow publishes core first.
