+++ .github/workflows/release.yml
@@ -37,12 +37,15 @@ jobs:
       - name: Tag matches the package versions
         run: |
           tag="${GITHUB_REF_NAME#v}"
-          for p in rn-iso expo-build-cache metro; do
+          for p in core rn-iso expo-build-cache metro; do
             v=$(node -p "require('./packages/$p/package.json').version")
             if [ "$v" != "$tag" ]; then
               echo "packages/$p is $v but the tag is $tag"; exit 1
             fi
           done
+      - name: Publish @rn-iso/core
+        run: npm publish --provenance --access public
+        working-directory: packages/core
       - name: Publish rn-iso
         run: npm publish --provenance --access public
         working-directory: packages/rn-iso
@@ -55,7 +58,7 @@ jobs:
       - name: Smoke-test the registry
         run: |
           sleep 10
-          for p in rn-iso @rn-iso/expo-build-cache @rn-iso/metro; do
+          for p in @rn-iso/core rn-iso @rn-iso/expo-build-cache @rn-iso/metro; do
             v=$(npm view "$p" version)
             echo "$p -> $v"
             [ "$v" = "${GITHUB_REF_NAME#v}" ] || exit 1
