+++ packages/metro/index.ts
@@ -26,20 +26,12 @@
 // Metro in-process both reach the published entry through require().
 
 import fs from 'node:fs';
-import os from 'node:os';
 import path from 'node:path';
+import { metroCacheRoot, registerCache } from '@rn-iso/core';
 
 // A Metro FileStore, kept structural so this package need not depend on
 // metro-cache's types: the constructor is all the caller relies on.
 type FileStoreCtor = new (options: { root: string }) => object;
-
-interface RegisterOptions {
-  dir: string;
-  name?: string;
-  prune?: string;
-  note?: string;
-  entriesDepth?: number;
-}
 
 // A Metro reporter event. Metro's event union is large and version-dependent, so
 // only the fields this reporter reads are named; the rest ride along untyped.
@@ -83,42 +75,8 @@ export interface NdjsonReporter {
 // quietly ignoring an override someone already set reads as an empty cache
 // rather than as an error. It names one directory, so it wins for a named cache
 // too -- otherwise half the stores on a machine would move and half would not.
-function configDir(): string {
-  return process.env.RN_ISO_HOME || path.join(os.homedir(), '.rn-iso');
-}
-
-// Anything that is not a plain path segment is replaced, and leading dots go, so
-// a scoped package name cannot climb out of the cache root.
-function cacheNameSegment(name: string): string {
-  return (
-    String(name)
-      .replace(/[^A-Za-z0-9._-]+/g, '-')
-      .replace(/^\.+/, '') || 'app'
-  );
-}
-
-// The machine config's override (`caches.metroCache` in <configDir>/config.json):
-// the same three-step resolution as the CLI's paths module -- env, config file,
-// default -- duplicated here because this package must work with no rn-iso
-// installed. Unreadable or malformed answers null; a cache override must never
-// be the reason a bundler config fails to load.
-function cachePathSetting(key: string): string | null {
-  try {
-    const parsed = JSON.parse(fs.readFileSync(path.join(configDir(), 'config.json'), 'utf-8')) as {
-      caches?: Record<string, unknown>;
-    };
-    const value = parsed?.caches?.[key];
-    return typeof value === 'string' && value.startsWith('/') ? value : null;
-  } catch {
-    return null;
-  }
-}
-
 export function cacheRoot(name?: string | null): string {
-  const setting = process.env.RN_ISO_METRO_CACHE || cachePathSetting('metroCache');
-  if (setting) return setting;
-  const root = path.join(configDir(), 'metro-cache');
-  return name === undefined || name === null || name === '' ? root : path.join(root, cacheNameSegment(name));
+  return metroCacheRoot(name);
 }
 
 // Registering makes this cache visible to `rn-iso gc`'s report, which is the
@@ -132,32 +90,6 @@ export function cacheRoot(name?: string | null): string {
 //   - rn-iso is an ES module, so `require` of it throws ERR_REQUIRE_ESM on Node
 //     before 20.19
 // A dynamic import fixes the second and not the first.
-function registerCache({ dir, name, prune, note, entriesDepth }: RegisterOptions): void {
-  try {
-    const home = configDir();
-    const file = path.join(home, 'caches.json');
-    let manifest: { version: number; caches: Array<Record<string, unknown>> } = { version: 1, caches: [] };
-    try {
-      const parsed = JSON.parse(fs.readFileSync(file, 'utf-8'));
-      if (Array.isArray(parsed?.caches)) manifest = { version: 1, caches: parsed.caches };
-    } catch {
-      // No manifest yet, or an unreadable one: start clean rather than fail.
-    }
-    // Keyed on the directory so repeated calls update rather than accumulate --
-    // these run on every build.
-    const others = manifest.caches.filter((c) => c.dir !== dir);
-    const record: Record<string, unknown> = { dir, name, prune, note, registeredBy: process.cwd() };
-    // Only written when the caller sets it: an absent depth means the entries
-    // are the directory's immediate children, which is the common case.
-    if (entriesDepth) record.entriesDepth = entriesDepth;
-    others.push(record);
-    fs.mkdirSync(home, { recursive: true });
-    fs.writeFileSync(file, JSON.stringify({ version: 1, caches: others }, null, 2));
-  } catch {
-    // A cache that cannot announce itself still works; it is just invisible.
-  }
-}
-
 function registerOnce(dir: string): void {
   registerCache({
     dir,
