+++ packages/rn-iso/package.json
@@ -42,6 +42,7 @@
     "build": "tsdown"
   },
   "dependencies": {
+    "@rn-iso/core": "^1.3.3",
     "@rn-iso/metro": "^1.2.0",
     "chalk": "^5.4.1",
     "commander": "^13.1.0"
